+++ formula/src/main/java/com/instacart/formula/RuntimeExtensions.kt
@@ -2,17 +2,17 @@ package com.instacart.formula
 
 import io.reactivex.rxjava3.core.Observable
 
-fun <State, RenderModel : Any> Formula<Unit, State, RenderModel>.start(): Observable<RenderModel> {
+fun <RenderModel : Any> IFormula<Unit, RenderModel>.start(): Observable<RenderModel> {
     return start(input = Unit)
 }
 
-fun <Input : Any, State, RenderModel : Any> Formula<Input, State, RenderModel>.start(
+fun <Input : Any, RenderModel : Any> IFormula<Input, RenderModel>.start(
     input: Input
 ): Observable<RenderModel> {
     return start(input = Observable.just(input))
 }
 
-fun <Input : Any, State, RenderModel : Any> Formula<Input, State, RenderModel>.start(
+fun <Input : Any, RenderModel : Any> IFormula<Input, RenderModel>.start(
     input: Observable<Input>
 ): Observable<RenderModel> {
     return FormulaRuntime.start(input = input, formula = this)